requirements.txt
streamlit 
agno 
pypdf
python-dotenv
groq
qdrant-client
openai
google-genai
